+++ resources/services/apps/replica_sets.go
@@ -18,7 +18,7 @@ func ReplicaSets() *schema.Table {
 		Resolver:     fetchAppsReplicaSets,
 		Multiplex:    client.ContextMultiplex,
 		DeleteFilter: client.DeleteContextFilter,
-		IgnoreError:  client.IgnoreForbidden,
+		IgnoreError:  client.IgnoreForbiddenNotFound,
 		Options:      schema.TableCreationOptions{PrimaryKeys: []string{"uid"}},
 		Columns: []schema.Column{
 			client.CommonContextField,
@@ -232,10 +232,10 @@ func ReplicaSets() *schema.Table {
 // ====================================================================================================================
 
 func fetchAppsReplicaSets(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
-	client := meta.(*client.Client).Services().ReplicaSets
+	cl := meta.(*client.Client).Services().ReplicaSets
 	opts := metav1.ListOptions{}
 	for {
-		result, err := client.List(ctx, opts)
+		result, err := cl.List(ctx, opts)
 		if err != nil {
 			return err
 		}