+++ resources/services/apps/stateful_sets.go
@@ -18,7 +18,7 @@ func StatefulSets() *schema.Table {
 		Resolver:     fetchAppsStatefulSets,
 		Multiplex:    client.ContextMultiplex,
 		DeleteFilter: client.DeleteContextFilter,
-		IgnoreError:  client.IgnoreForbidden,
+		IgnoreError:  client.IgnoreForbiddenNotFound,
 		Options:      schema.TableCreationOptions{PrimaryKeys: []string{"uid"}},
 		Columns: []schema.Column{
 			client.CommonContextField,
@@ -292,10 +292,10 @@ func StatefulSets() *schema.Table {
 // ====================================================================================================================
 
 func fetchAppsStatefulSets(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
-	client := meta.(*client.Client).Services().StatefulSets
+	cl := meta.(*client.Client).Services().StatefulSets
 	opts := metav1.ListOptions{}
 	for {
-		result, err := client.List(ctx, opts)
+		result, err := cl.List(ctx, opts)
 		if err != nil {
 			return err
 		}