+++ resources/services/core/endpoints.go
@@ -7,10 +7,9 @@ import (
 	"net"
 
 	"github.com/cloudquery/cq-provider-k8s/client"
+	"github.com/cloudquery/cq-provider-sdk/provider/schema"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"github.com/cloudquery/cq-provider-sdk/provider/schema"
 )
 
 func Endpoints() *schema.Table {
@@ -20,7 +19,7 @@ func Endpoints() *schema.Table {
 		Resolver:     fetchCoreEndpoints,
 		Multiplex:    client.ContextMultiplex,
 		DeleteFilter: client.DeleteContextFilter,
-		IgnoreError:  client.IgnoreForbidden,
+		IgnoreError:  client.IgnoreForbiddenNotFound,
 		Options:      schema.TableCreationOptions{PrimaryKeys: []string{"uid"}},
 		Columns: []schema.Column{
 			client.CommonContextField,